+++ src/testmod/java/net/frozenblock/lib/testmod/mixin/SonicBoomMixin.java
@@ -17,7 +17,8 @@
 
 package net.frozenblock.lib.testmod.mixin;
 
-import net.frozenblock.lib.screenshake.api.ScreenShakeManager;
+import net.frozenblock.lib.screenshake.api.ScreenShake;
+import net.frozenblock.lib.screenshake.api.ScreenShakes;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.behavior.warden.SonicBoom;
@@ -30,8 +31,16 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(SonicBoom.class)
 public class SonicBoomMixin {
 
-	@Inject(method = {"m_ehrxwrfs","method_43265","lambda$tick$2"}, at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/monster/warden/Warden;playSound(Lnet/minecraft/sounds/SoundEvent;FF)V", ordinal = 0), require = 1)
-	private static void startShaking(Warden warden, ServerLevel world, LivingEntity livingEntity, CallbackInfo ci) {
-		ScreenShakeManager.addScreenShake(world, 0.5F, 25, warden.getX(), warden.getY(), warden.getZ(), 18);
+	@Inject(
+		method = "lambda$tick$2",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/entity/monster/warden/Warden;playSound(Lnet/minecraft/sounds/SoundEvent;FF)V",
+			ordinal = 0
+		),
+		require = 1
+	)
+	private static void startShaking(Warden body, ServerLevel level, LivingEntity target, CallbackInfo info) {
+		ScreenShakes.addScreenShake(body, ScreenShake.builder(body).intensity(0.5F).duration(25).maxDistance(18).build());
 	}
 }